+++ test/Faucet.t.sol
@@ -11,40 +11,26 @@ import "forge-std/console.sol";
 contract FaucetTest is Test {
     using ECDSA for bytes32;
     using SignatureChecker for address;
+
     Faucet _faucet;
 
     address _owner = address(1);
     address _user = vm.addr(3);
 
     // Create a aux function to create a signature for claiming kinto ETH from the faucet
-    function _auxCreateSignature(address _signer, uint256 _privateKey, uint256 _expiresAt) private view returns (
-        IFaucet.SignatureData memory signData
-    ) {
+    function _auxCreateSignature(address _signer, uint256 _privateKey, uint256 _expiresAt)
+        private
+        view
+        returns (IFaucet.SignatureData memory signData)
+    {
         bytes32 dataHash = keccak256(
-            abi.encode(
-                _signer,
-                address(_faucet),
-                _expiresAt,
-                _faucet.nonces(_signer),
-                bytes32(block.chainid)
-            ));
-        bytes32 hash = keccak256(
-            abi.encodePacked(
-                bytes1(0x19),
-                bytes1(0x01),
-                dataHash
-            )
-        ).toEthSignedMessageHash();
+            abi.encode(_signer, address(_faucet), _expiresAt, _faucet.nonces(_signer), bytes32(block.chainid))
+        );
+        bytes32 hash = keccak256(abi.encodePacked(bytes1(0x19), bytes1(0x01), dataHash)).toEthSignedMessageHash();
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(_privateKey, hash);
         bytes memory signature = abi.encodePacked(r, s, v);
-        return IFaucet.SignatureData(
-                _signer,
-                _faucet.nonces(_signer),
-                _expiresAt,
-                signature
-            );
+        return IFaucet.SignatureData(_signer, _faucet.nonces(_signer), _expiresAt, signature);
     }
-
 
     function setUp() public {
         vm.chainId(1);
@@ -85,7 +71,7 @@ contract FaucetTest is Test {
         _faucet.startFaucet{value: 1 ether}();
         vm.stopPrank();
         vm.startPrank(_user);
-        uint previousBalance = address(_user).balance;
+        uint256 previousBalance = address(_user).balance;
         _faucet.claimKintoETH();
         assertEq(address(_faucet).balance, 1 ether - _faucet.CLAIM_AMOUNT());
         assertEq(address(_user).balance, previousBalance + _faucet.CLAIM_AMOUNT());
